Create Regression of Insurance Data
This is a multiple linear regression model in python that takes six variables/risk factors and uses them to predict cost ("charges").  The data to train the model can be found here: https://www.kaggle.com/datasets/willianoliveiragibin/healthcare-insurance.

Explanations for each piece of the code are in #hashes preceding the syntax.

At the end there is a function that creates different scatterplots from matplotlib as well as a function that gives some statistics if you want to filter the data.  For ex/ mean "charges" (cost) for individuals who are age: 27 or mean age for individuals who have 5 children.